requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
crispy-tailwind     1.0.3
Django              5.1.1
django-crispy-forms 2.3
pillow              10.4.0
pip                 24.2
sqlparse            0.5.1
tzdata              2024.2
